Move functions to pyanaconda.core.storage
Move functions from pyanaconda.storage.utils that are used by the Storage module
and UI to pyanaconda.core.storage
