+++ pyanaconda/modules/storage/partitioning/custom/custom_partitioning.py
@@ -36,7 +36,8 @@ from pyanaconda.modules.storage.partitioning.automatic.noninteractive_partitioni
 from pyanaconda.modules.storage.partitioning.automatic.utils import get_candidate_disks, \
     schedule_partitions
 from pyanaconda.platform import platform
-from pyanaconda.storage.utils import suggest_swap_size, get_pbkdf_args, lookup_alias
+from pyanaconda.storage.utils import get_pbkdf_args, lookup_alias
+from pyanaconda.core.storage import suggest_swap_size
 
 log = get_module_logger(__name__)
