+++ pyanaconda/ui/gui/spokes/lib/accordion.py
@@ -21,7 +21,7 @@ from pyanaconda.anaconda_loggers import get_module_logger
 from pyanaconda.core.constants import DEFAULT_AUTOPART_TYPE
 from pyanaconda.core.i18n import _, C_
 from pyanaconda.product import productName, productVersion
-from pyanaconda.storage.utils import get_supported_autopart_choices
+from pyanaconda.core.storage import get_supported_autopart_choices
 from pyanaconda.ui.gui.utils import escape_markup, really_hide, really_show
 
 import gi
